Merg my js files into index.js
